requirements.txt
qdrant-client
PyMuPDF
langchain
sentence-transformers
torch
transformers
Pillow
openai
python-dotenv
tqdm
fastapi
uvicorn
